requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Database connectors
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Streaming and messaging
kafka-python>=2.0.0
confluent-kafka>=2.0.0

# Big Data processing
pyspark>=3.4.0
delta-spark>=2.4.0

# Workflow orchestration
apache-airflow>=2.7.0
apache-airflow-providers-postgres>=5.0.0
apache-airflow-providers-apache-kafka>=3.0.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
prophet>=1.1.0

# Visualization and reporting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
jupyter>=1.0.0
ipykernel>=6.25.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools
black>=23.0.0
flake8>=6.0.0
pre-commit>=3.3.0 